Add magic_calculation function to perform calculations
This commit adds a new function called magic_calculation to perform calculations based on the values of 'a' and 'b'. The function imports the add and sub functions from the magic_calculation_102 module and uses them to perform the calculations. If 'a' is less than 'b', the function adds 'a' and 'b' and then adds the values of 'y' in the range of 4 to 6 to the result. If 'a' is greater than or equal to 'b', the function subtracts 'b' from 'a'. This commit introduces the necessary logic to perform the calculations.